Split FireDb op i flere filer
Med henblik på at øge overskueligheden af FireDb-klassen splittes
den op i flere filer. For nuværende er tre grupper er metoder flyttet
til seperate moduler: Hent, indsæt og luk. Skulle der i fremtiden
opstå flere naturlige grupperinger, fx søg eller tilknyt, bør disse
også udskilles i seperate moduler.
